Adjust behavior of set_attrs on solaris
set_attrs is expecting to set file times on symlinks.  The original
function falls back to utimes which is a bug (utimes always follows sym
links and tries to update times there, even if file does not exist).
However, the bug isn't known because all supported platforms support
utimensat (even latest solaris 10, but not back in u3)

So on solaris, check if the target file is a symlink.  if so, skip it.